Use remix custom routes
This allows a more sensible, domain-driven routing. Just add `.route` to
the file. Et voila!
