+++ app/recipes/recipes.$recipeId.route.tsx
@@ -8,13 +8,13 @@ import {
 } from "@remix-run/react";
 import invariant from "tiny-invariant";
 
-import { loadSingleRecipe } from "~/api/recipe-loader";
+import { loadSingleRecipe } from "~/recipes/recipe-loader";
 import { List } from "~/components/lists";
 import { Time } from "~/components/time";
 import { useKeyboardCombo } from "~/components/use-keyboard";
 import { getComments } from "~/models/comment.server";
 import { deleteRecipe } from "~/models/recipe.server";
-import { CommentListAndForm, isFlatComment } from "~/routes/api.comments";
+import { CommentListAndForm, isFlatComment } from "~/routes/api.comments.route";
 import { requireUserId } from "~/session.server";
 
 export const loader = async (args: LoaderFunctionArgs) => {
