+++ app/routes/api.restricted.route.tsx
