+++ app/routes/explore.route.tsx
