+++ app/routes/healthcheck.route.tsx
