+++ app/routes/logout.route.tsx
