+++ package.json
@@ -46,6 +46,7 @@
     "isbot": "^3.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "remix-custom-routes": "^1.0.1",
     "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
